Build trigger:  python all

Binaries submitted in 962896
Build time : 35 minutes 12.490000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/blue/": change = 962870]

+++ Include/BitPacker.h
@@ -942,7 +942,7 @@ bool BitPackerCore::Unpack( wchar_t *string, unsigned int *len, unsigned int max
 		*len = tlen;
 	}
 		
-	return (tlen != (size_t)-1) && Valid();
+	return (tlen != (unsigned int)-1) && Valid();
 }
 
 //------------------------------------------------------------------------------
